requirements.txt
yfinance>=0.2.31
pandas>=2.0.0
requests>=2.31.0
anthropic>=0.39.0
python-dotenv>=1.0.0
